refactor: replace Double with BigDecimal for cost calculations and update BillOfMaterial structure with removing the strong coupling between production module and supply module with adding an facade interface

+++ src/main/java/net/ayman/supplychainx/production/dto/bom/BillOfMaterialResponseDTO.java
@@ -2,6 +2,8 @@ package net.ayman.supplychainx.production.dto.bom;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BillOfMaterialResponseDTO {
     private Long id;
@@ -11,6 +13,6 @@ public class BillOfMaterialResponseDTO {
     private String productName;
     private String materialUnit;
     private Integer quantity;
-    private Double unitCost;
-    private Double totalCost;
+    private BigDecimal unitCost;
+    private BigDecimal totalCost;
 }
